requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
torch==2.1.0
transformers==4.35.0
pillow==10.1.0
numpy<2
python-multipart==0.0.6
tensorflow==2.15.0

